+++ docs/source/changelog.rst
@@ -16,6 +16,11 @@ Core
 
 * Added :class:`KinBody.LinkInfo` and :class:`KinBody.JointInfo` classes in order to programatically build robots by calling :meth:`.KinBody.Init`.
 
+Misc
+----
+
+* Fixed :meth:`.planningutils.JitterTransform` and added openravepy wrapper
+
 Version 0.8.2
 =============